requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.8
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2